+++ src/main/kotlin/observer/lifeCycle/LifeCycleOwner.kt
@@ -1,3 +1,3 @@
-package lifeCycle
+package observer.lifeCycle
 
 class LifeCycleOwner(val lifeCycle: LifeCycle)
