+++ src/test/kotlin/DataObserverTest.kt
@@ -1,4 +1,4 @@
-import observer.DataObserver
+import observer.observer.DataObserver
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
